Custom folder small bug fix
I get this message when hunting-buddy tries to run one of the scripts in the custom folder.
Same when I try to use DRC.wait_for_script_to_complete().

```R>;e verify_script('CC')
--- Lich: exec1 active.
[exec1: Failed to find a script named 'CC']
[exec1: Please report this to <https://github.com/rpherbig/dr-scripts/issues>]
[exec1: or to Discord <https://discord.gg/ffcEAYfK>]
--- Lich: exec1 has exited.```

I didn't use the File.join() method for these because it caused problems for all of the ones with the extensions included.  I just matched the existing code instead.