Custom script already running is not detected
The check to see if a script is already running when you try to run a script isn't detecting scripts from the custom folder running.  This fixes that issue.

The problem is that it's comparing the `script_name` for the script that the user wants to run against the names from `Script.running + Script.hidden` and the `script_name` includes `/custom/` at the front for any scripts in the custom folder.

I put in a sub to remove `/custom/ from the `script_name` when it compares the new script to already running scripts.

I also added to the message that gets displayed when a script is not found.  Previously it only showed that it was not found in the `./scripts` directory.  I added `./scripts/custom` to that message to reflect that we're searching both places.